requirements.txt
# ── Core scientific stack ─────────────────────────────────────
numpy>=1.26.0
pandas>=2.1.0
scipy>=1.11.0
scikit-learn>=1.3.0

# ── Data acquisition ─────────────────────────────────────────
yfinance>=0.2.35
fredapi>=0.5.1

# ── Time series / financial models ───────────────────────────
statsmodels>=0.14.0
arch>=6.2.0          # GARCH volatility models
hmmlearn>=0.3.2      # Hidden Markov Models

# ── Portfolio analytics & tear sheets ────────────────────────
quantstats>=0.0.62

# ── Visualization ─────────────────────────────────────────────
matplotlib>=3.7.0
seaborn>=0.13.0
plotly>=5.17.0

# ── NLP / Sentiment (optional — FinBERT) ─────────────────────
# Uncomment to enable sentiment signal:
# transformers>=4.36.0
# torch>=2.1.0

# ── Config & utilities ───────────────────────────────────────
pyyaml>=6.0
python-dotenv>=1.0.0  # for env var override of API keys

# ── Notebook environment ─────────────────────────────────────
jupyter>=1.0.0
ipykernel>=6.0.0
ipywidgets>=8.0.0

# ── Testing ───────────────────────────────────────────────────
pytest>=7.4.0
